requirements.txt
fastapi
uvicorn
python-dotenv
openai
anthropic
google-generativeai
chromadb
beautifulsoup4
requests
fake-useragent
pydantic
sentence-transformers
